Branch that uses floating point for texture coordinates
For platforms like Dreamcast which don't support integer texture coordinates

+++ rvm/Core/DrawVertex.h
@@ -8,8 +8,8 @@
 #include "SDL.h"
 
 struct Vector2 {
-    short X;
-    short Y;
+    float X;
+    float Y;
 };
 
 struct Color {
